tiny_yolov2/requirements.txt
attrdict
Brotli
certifi
cffi
gevent
geventhttpclient
greenlet
grpcio
opencv-python
numpy
Pillow
protobuf
pycparser
python-rapidjson
six
tritonclient[all]
zope.event
zope.interface
wget